Rename DebugPanel to Panel
I'm seeing lots of DeprecationWarnings in my test logs, I think it's down to DDT renaming their base panel class.

I've changed the name here but I've not tested it or anything. I'm assuming it's a drop in replacement. If this isn't the case I'm more than happy to put some actual time into a PR. This was just made using GitHub's in-browser editor.